sync

+++ cli/dotnet/Logic/Repository/Models/BookEntity.cs
@@ -2,13 +2,13 @@ namespace Exercises.Logic.Repository.Models;
 
 public class BookEntity
 {
-    public int Id { get; set; }
+    public long Id { get; set; }
     public string Title { get; set; }
     public string Authors { get; set; }
-    public int PageStart { get; set; }
-    public int PageEnd { get; set; }
+    public long PageStart { get; set; }
+    public long PageEnd { get; set; }
     public string Reference { get; set; }
-    public int TopicId { get; set; }
+    public long TopicId { get; set; }
     public TopicEntity Topic { get; set; }
 
     public List<ChapterEntity> Chapters { get; set; } = [];
